requirements.txt
Flask==3.1.2
Flask-WTF==1.2.2
mysql-connector-python==9.5.0
WTForms==3.2.1
email-validator==2.3.0
